requirements.txt
GitPython
requests
geopy
ip2geotools
emoji
telethon
pysocks
tqdm
pybase64
filelock
python-dotenv